pratica - redes/requirements.txt
colorama==0.4.6
pycryptodome==3.20.0
pyotp==2.9.0
pypng==0.20220715.0
qrcode==7.4.2
typing_extensions==4.11.0
